requirements.txt
# ── Core ────────────────────────────────────────────────────────────
flask>=3.0
python-dotenv>=1.0
pandas>=2.0
openpyxl>=3.1       # lectura de Excel (.xlsx)
xlsxwriter>=3.1     # escritura de Excel con formatos

# ── Motores de base de datos (instala solo el que uses) ──────────────

# SQL Server — Windows o Linux con ODBC Driver 17/18
pyodbc>=5.0

# PostgreSQL
# psycopg2-binary>=2.9   # descomenta si usas PostgreSQL

# MySQL / MariaDB
# PyMySQL>=1.1           # descomenta si usas MySQL

# Seguridad — hashing de contraseñas (incluido con Flask, se lista explícitamente)
werkzeug>=3.0
